fix(perception): store eventgen clock as i128 and wrap RGB admission as ContentDigest [fss-ub4yv]
TimestampNs is i128; keep clock comparable without truncation.
RgbFrameAdmission now takes ContentDigest (sha256 of admission bytes)
instead of a raw digest array on encode/decode/verify.

+++ crates/fss-reference/src/ingest/eventgen.rs
@@ -124,7 +124,7 @@ pub struct ZoneEventGenerator {
     last_emitted: HashMap<(String, u64), TimestampNs>,
     /// Highest accepted emission time; observations below it are refused
     /// rather than silently absorbed into a cooldown window.
-    clock: Option<i64>,
+    clock: Option<i128>,
     event_seq: u64,
 }
